requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
cssselect2==0.7.0
Django==4.2.2
django-ckeditor==6.6.1
django-crispy-forms==2.0
django-embed-video==1.4.8
django-js-asset==2.0.0
django-widget-tweaks==1.4.8
gunicorn==21.2.0
html5lib==1.1
idna==3.4
lxml==4.9.2
mysqlclient==2.1.1
oscrypto==1.3.0
packaging==23.1
Pillow==9.4.0
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.9.1
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2021.1
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.3.1
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.6.3
tzdata==2022.7
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.3
webencodings==0.5.1
xhtml2pdf==0.2.11
